requirements.txt
numpy
scipy
matplotlib
torch
torchvision
tensorboard
pytorch-lightning
notebook
python-box 
PyYAML
